+++ pos-server/src/main/java/com/saulpos/server/security/repository/RoleRepository.java
@@ -3,9 +3,12 @@ package com.saulpos.server.security.repository;
 import com.saulpos.server.security.model.RoleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
 import java.util.Optional;
 
 public interface RoleRepository extends JpaRepository<RoleEntity, Long> {
 
     Optional<RoleEntity> findByCode(String code);
+
+    List<RoleEntity> findAllByOrderByCodeAsc();
 }